# Sidebar Render Improvements