Summary of changes:

* Adds general method  to base view for bulk appending of views, with only one repaint per bulk append.
* Implements method for sidebar entry views
* Implements method for topic node inner views on path navigation

@66eli77 If you could benchmark this against current develop and post the results here, I would be much obliged!