@66eli77 2015-04-24
my pleasure, will perform the benchmark later today once I fix a android bug I'm working on