@66eli77 2015-04-28
![screen shot 2015-04-28 at 10 23 23 am](https://cloud.githubusercontent.com/assets/5577205/7375915/b307da06-ed90-11e4-960f-5b44a198f316.png)
new change on the left / old version on the right