@rtibbles 2015-04-28
Do we have the capability to run benchmarks multiple times and take averages to get more robust estimates of these? I can't think why loading would be particularly different in either case.