@66eli77 2015-04-30
yes, that feature request has been implemented, and here's the result:
Old: Avg Painting Time: 107.590133333 ms
New: Avg Painting Time: 106.897466667 ms
(these are the avg. value of 30 runs each)