@rtibbles 2015-04-30
@66eli77 this has now been rebased to latest develop, where the sidebar icons have been reinstated. When you have time, a repeat of the benchmark would be much appreciated!