@66eli77 2015-05-01
I just pulled from the latest develop, but he icons are still missing