@rtibbles 2015-05-01
Odd. Will double check.

On Fri, 1 May 2015 14:10 Eli <notifications@github.com> wrote:

> I just pulled from the latest develop, but he icons are still missing
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/learningequality/ka-lite/pull/3555#issuecomment-98239675>
> .
>