@rtibbles 2015-05-12
@66eli77 is re-benchmarking it soon.