@66eli77 2015-05-13
benchmark with icons on the latest develop branch:
original:
Run 1 : 101.127 ms
Run 2 : 86.362 ms
Run 3 : 92.352 ms
Run 4 : 95.174 ms
Run 5 : 93.18 ms
Run 6 : 103.037 ms
Run 7 : 115.511 ms
Run 8 : 92.099 ms
Run 9 : 96.843 ms
Run 10 : 144.623 ms
Run 11 : 144.505 ms
Run 12 : 127.826 ms
Run 13 : 119.603 ms
Run 14 : 121.582 ms
Run 15 : 112.737 ms
Run 16 : 108.168 ms
Run 17 : 150.58 ms
Run 18 : 143.251 ms
Run 19 : 147.286 ms
Run 20 : 134.74 ms
Run 21 : 137.811 ms
Run 22 : 132.937 ms
Run 23 : 123.505 ms
Run 24 : 138.214 ms
Run 25 : 144.084 ms
Run 26 : 145.28 ms
Run 27 : 162.78 ms
Run 28 : 146.952 ms
Run 29 : 129.197 ms
Run 30 : 165.568 ms
Avg Painting Time: 125.230466667 ms

Richard:
Run 1 : 137.784 ms
Run 2 : 114.425 ms
Run 3 : 158.712 ms
Run 4 : 146.193 ms
Run 5 : 153.206 ms
Run 6 : 158.612 ms
Run 7 : 145.478 ms
Run 8 : 138.585 ms
Run 9 : 137.728 ms
Run 10 : 123.354 ms
Run 11 : 136.352 ms
Run 12 : 140.812 ms
Run 13 : 131.857 ms
Run 14 : 146.566 ms
Run 15 : 137.755 ms
Run 16 : 142.189 ms
Run 17 : 152.489 ms
Run 18 : 132.969 ms
Run 19 : 125.519 ms
Run 20 : 134.374 ms
Run 21 : 153.55 ms
Run 22 : 133.586 ms
Run 23 : 141.082 ms
Run 24 : 138.813 ms
Run 25 : 143.324 ms
Run 26 : 142.949 ms
Run 27 : 127.19 ms
Run 28 : 119.154 ms
Run 29 : 135.407 ms
Run 30 : 138.961 ms
Avg Painting Time: 138.965833333 ms