@rtibbles 2015-05-13
Doesn't look like this is giving any performance benefits at all.

It seems, therefore, that our poor render performance is not due to multiple reflows of the DOM, as that is exactly what this should be preventing.

@66eli77 Do you have any insights into what is causing slow performance if it is not this?