@66eli77 2015-05-14
I looked at the timeline graph, the paint pattern looks exactly the same, therefore, the new changes neither got applied or function as we expected