@66eli77 2015-05-14
I think this stackoverflow thread explained the exact problem we have here: http://stackoverflow.com/questions/14203196/does-using-a-document-fragment-really-improve-performance