@rtibbles 2015-05-14
Fair enough, seems like in modern browsers, this kind of optimization is
being done under the hood already.

So, what is causing the performance problems for the sidebar?

On Thu, 14 May 2015 at 14:16 Eli <notifications@github.com> wrote:

> I think this stackoverflow thread explained the exact problem we have
> here:
> http://stackoverflow.com/questions/14203196/does-using-a-document-fragment-really-improve-performance
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/learningequality/ka-lite/pull/3555#issuecomment-102171391>
> .
>