@rtibbles 2015-05-18
This gives no apparent improvements. Closing.